refactor: request, response DTO 이름 및 패키지 구조 변경
