--- BE/src/main/java/codesquad/backend/issuetracker/label/presentation/dto/LabelCountDto.java
@@ -1,9 +1,0 @@
-package codesquad.backend.issuetracker.label.presentation.dto;
-
-import lombok.Getter;
-
-@Getter
-public class LabelCountDto {
-
-	private Long labelCount;
-}
