--- BE/src/main/java/codesquad/backend/issuetracker/label/presentation/dto/LabelsResponseDto.java
@@ -1,10 +1,0 @@
-package codesquad.backend.issuetracker.label.presentation.dto;
-
-import java.util.List;
-import lombok.Getter;
-
-@Getter
-public class LabelsResponseDto {
-
-	private List<LabelDto> labels;
-}
