+++ BE/src/main/java/codesquad/backend/issuetracker/label/presentation/dto/response/LabelCountResponse.java
@@ -1,0 +1,9 @@
+package codesquad.backend.issuetracker.label.presentation.dto.response;
+
+import lombok.Getter;
+
+@Getter
+public class LabelCountResponse {
+
+	private Long labelCount;
+}
