+++ BE/src/main/java/codesquad/backend/issuetracker/label/presentation/dto/response/LabelsResponse.java
@@ -1,0 +1,11 @@
+package codesquad.backend.issuetracker.label.presentation.dto.response;
+
+import codesquad.backend.issuetracker.label.presentation.dto.LabelDto;
+import java.util.List;
+import lombok.Getter;
+
+@Getter
+public class LabelsResponse {
+
+	private List<LabelDto> labels;
+}
